update job detail url
